requirements.txt
greenlet==1.1.0
pika==1.2.0
pika-stubs==0.1.3
psycopg2-binary==2.9.1
python-dotenv==0.19.0
SQLAlchemy==1.4.22
